+++ tests/test_version.py
@@ -53,7 +53,8 @@ class TestVersionModule(unittest.TestCase):
             or cur.startswith("2.4.")
             or cur.startswith("2.5.")
             or cur.startswith("2.6.")
-            or cur.startswith("2.7."),
+            or cur.startswith("2.7.")
+            or cur.startswith("2.8."),
             cur,
         )
         if "rc" in cur.lower():